dom/tests/mochitest/bugs/test_bug479143.html

<!DOCTYPE html>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=479143
-->
<head>
  <title>Test for Bug 411103</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=479143">Mozilla Bug 479143</a>
<p id="display"></p>
<div id="content" style="display: none"></div>

<pre id="test">
<script class="testbody" type="text/javascript">

  if (!navigator.platform.startsWith("Win")) {
    SimpleTest.expectAssertions(1);
  }

  SimpleTest.waitForExplicitFinish();

  setTimeout(function() {
    var interval = setInterval(function() { var i = 0; i++; }, 10);

    var xhr = new XMLHttpRequest();
    xhr.open("GET", "test_bug479143.html", false);
    xhr.send(null);

    window.showModalDialog("javascript:" +
                           "setTimeout(function() { window.close(); }, 1000);",
                           null);

    clearInterval(interval);

    ok(true, "did not crash");

    // Garbage collecting the windows created in this test can cause
    // assertions, so GC now to blame those assertions to this test.
    // ("mArguments wasn't cleaned up properly!" in ~nsGlobalWindow,
    // bug 600703)
    SpecialPowers.gc();

    SimpleTest.finish();
  }, 0);

</script>
</pre>
</body>
</html>
